refactor: extract RouterCollection to a separate file
